Finish the upgrade to the new Equations, Solutions and Makie interfaces
Completes the work in progress on the guiding-centre-3d branch.

Dependencies and package layout:
* move plotting into a Makie + LaTeXStrings package extension and drop the
  Plots.jl, RecipesBase, GeometricProblems, GeometricIntegrators(Base) and
  ForwardDiff dependencies
* drop the unused Documenter, EulerLagrange and Printf dependencies and the
  stale SimpleSolvers test extra
* move the initial conditions and coordinate helpers to src/utils
* raise the minimum Julia version to 1.10

Guiding centre dynamics in 3D:
* add the canonicalised formulation, hode_canonical, whose Hamiltonian absorbs
  the Lagrange multipliers of the two constraints, and include it from all
  thirteen equilibria
* add a documentation page for the model

Diagnostics:
* port compute_energy, compute_toroidal_momentum and their error variants from
  GeometricProblems.Diagnostics to GeometricSolutions, threading the parameters
  through the new compute_invariant interface
* re-export compute_momentum_error, compute_one_form and compute_error_drift
  from GeometricSolutions, which were exported but no longer defined
* replace convert_coordinates_RZphi_to_xyz, which used the removed SDataSeries,
  by cartesian_solution

Plotting:
* port the poloidal, 3D, component and Poincaré invariant plots to Makie as
  plain functions returning figures, replacing the Plots.jl recipes and the
  PyPlot code
* plot_trajectory_3d_figure declared positional instead of keyword arguments
* plot_invariant_error divided by log10(0) for a vanishing error series

Field modules:
* toroidal_momentum referred to the undefined ϑ3 instead of ϑ₃ in four 3D and
  four 4D equilibria
* tokamak_medium_cartesian exported toroidal_momentum without defining it

Tests:
* the guiding centre 4D tests imported MidpointExtrapolation from
  GeometricIntegratorsBase, which is no longer available in the test
  environment, so the suite never ran past that point
* cover hode_canonical for the active 3D guiding centre equilibria
* reduce the time step where the solvers did not converge at the equilibrium
  defaults, and use midpoint extrapolation for the Pauli theta pinch, whose
  constant momentum makes the default Hermite extrapolation degenerate

Documentation:
* reactivate the initial conditions figure using Makie
* fix the example block of the ITER magnetic field, which was never executed
