+++ Project.toml
@@ -4,45 +4,35 @@ version = "0.1.0"
 authors = ["Michael Kraus <michael.kraus@ipp.mpg.de>"]
 
 [deps]
-Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 ElectromagneticFields = "d6c1ba6f-ee03-53af-b876-68cefeb88ec8"
-EulerLagrange = "e4b6a126-98cc-486b-a68d-c3e5d57f8cc9"
-ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
 GeometricEquations = "c85262ba-a08a-430a-b926-d29770767bf2"
-GeometricIntegrators = "dcce2d33-59f6-5b8d-9047-0defad88ae06"
-GeometricIntegratorsBase = "71212ab4-08a2-48eb-a02b-4cbad0e9c44a"
-GeometricProblems = "18cb22b4-ad41-5c80-9c5f-710df63fbdc9"
 GeometricSolutions = "7843afe4-64f4-4df4-9231-049495c56661"
-LaTeXStrings = "b964fa9f-0449-5b57-a5c2-d3ea65f4040f"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 PoincareInvariants = "26663084-47d3-540f-bd97-40ca743aafa4"
-Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
-RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
+
+[weakdeps]
+LaTeXStrings = "b964fa9f-0449-5b57-a5c2-d3ea65f4040f"
+Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
+
+[extensions]
+ChargedParticlePlots = ["Makie", "LaTeXStrings"]
 
 [compat]
-Documenter = "1"
 ElectromagneticFields = "0.6"
-EulerLagrange = "0.4.4"
-ForwardDiff = "0.10, 1"
 GeometricEquations = "0.20"
 GeometricIntegrators = "0.15.1"
-GeometricIntegratorsBase = "0.1.5"
-GeometricProblems = "0.6"
-GeometricSolutions = "0.5"
+GeometricSolutions = "0.5.9"
 LaTeXStrings = "1"
+Makie = "0.24"
 Parameters = "0.12, 0.13"
-Plots = "1"
 PoincareInvariants = "0.4"
-RecipesBase = "1"
-julia = "1.6"
+julia = "1.10"
 
 [extras]
 GeometricIntegrators = "dcce2d33-59f6-5b8d-9047-0defad88ae06"
 SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
-SimpleSolvers = "36b790f5-6434-4fbe-b711-1f64a1e2f6a2"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["GeometricIntegrators", "SafeTestsets", "SimpleSolvers", "Test"]
+test = ["GeometricIntegrators", "SafeTestsets", "Test"]
