+++ docs/Makefile
@@ -2,7 +2,7 @@
 all: documenter
 
 documenter:
-	julia --color=yes --project=.. make.jl
+	julia --color=yes --project=. make.jl
 
 clean:
 	rm -Rf build
