+++ docs/Project.toml
@@ -1,7 +1,12 @@
 [deps]
+CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
 ChargedParticleDynamics = "209b3593-7183-5600-9f8f-df5ce7fddd64"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 ElectromagneticFields = "d6c1ba6f-ee03-53af-b876-68cefeb88ec8"
 GeometricIntegrators = "dcce2d33-59f6-5b8d-9047-0defad88ae06"
 LaTeXStrings = "b964fa9f-0449-5b57-a5c2-d3ea65f4040f"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+Markdown = "d6f4376e-aef5-505a-96c1-9c027394607a"
+
+[sources]
+ChargedParticleDynamics = {path = ".."}
