+++ docs/make.jl
@@ -6,13 +6,19 @@ makedocs(
     warnonly = Documenter.except(:autodocs_block, :cross_references, :docs_block, :doctest, :eval_block, :example_block, :footnote, :linkcheck_remotes, :linkcheck, :meta_block, :parse_error, :setup_block),
     format = Documenter.HTML(
                 prettyurls = get(ENV, "CI", nothing) == "true",
-                assets = [asset("assets/style.css", class=:css, islocal=true)]),
+                assets = [asset("assets/style.css", class=:css, islocal=true)],
+                example_size_threshold = 32,
+                ),
     pages = ["Overview"                      => "index.md",
              "Normalization"                 => "normalization.md",
              "Initialization"                => "initialization.md",
              "Charged Particles in 3D"       => "charged_particle_3d.md",
              "Pauli Particles in 3D"         => "pauli_particle_3d.md",
+             "Guiding Center Dynamics in 3D" => "guiding_center_3d.md",
              "Guiding Center Dynamics in 4D" => "guiding_center_4d.md",
+             "Examples" => [
+                    "ITER Equilibrium im Cylindrical Coordinates"   => "examples/iter_cylindrical.md",
+                 ],
             ]
 )
