+++ examples/Project.toml
@@ -3,8 +3,8 @@ CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
 ChargedParticleDynamics = "209b3593-7183-5600-9f8f-df5ce7fddd64"
 DecFP = "55939f99-70c6-5e9b-8bb0-5071ed7d61fd"
 ElectromagneticFields = "d6c1ba6f-ee03-53af-b876-68cefeb88ec8"
+GLMakie = "e9467ef8-e4e7-5192-8a1a-b1aee30e663a"
 GeometricIntegrators = "dcce2d33-59f6-5b8d-9047-0defad88ae06"
-GeometricIntegratorsBase = "71212ab4-08a2-48eb-a02b-4cbad0e9c44a"
 GeometricProblems = "18cb22b4-ad41-5c80-9c5f-710df63fbdc9"
 LaTeXStrings = "b964fa9f-0449-5b57-a5c2-d3ea65f4040f"
 Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
@@ -13,3 +13,6 @@ PoincareInvariants = "26663084-47d3-540f-bd97-40ca743aafa4"
 Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
 SimpleSolvers = "36b790f5-6434-4fbe-b711-1f64a1e2f6a2"
 Weave = "44d3d7a6-8a23-5bf8-98c5-b353f8df5ec9"
+
+[sources]
+ChargedParticleDynamics = {path = ".."}
