+++ src/ChargedParticleDynamics.jl
@@ -1,12 +1,14 @@
 module ChargedParticleDynamics
 
-    include("initialization/initial_conditions.jl")
+    include("utils/coordinates.jl")
+    include("utils/initial_conditions.jl")
 
     include("ChargedParticle3d.jl")
     include("GuidingCenter3d.jl")
     include("GuidingCenter4d.jl")
     include("GyroKinetics4d.jl")
     include("PauliParticle3d.jl")
-    include("ChargedParticlePlots.jl")
+
+    include("Plots.jl")
 
 end
