+++ src/guiding_center_3d/dipole.jl
@@ -11,13 +11,13 @@ export hamiltonian
 
 Dipole.@code() # inject magnetic field code
 
-const Δt = 0.035
-const tspan = (0.0, 35.0)
-# const tspan = (0.0, 999.99)
+const Δt = 0.03
+const tspan = (0.0, 90_000.0)
 
 initial_conditions_dipole() = (initial_conditions(0.0, [1.0, 2.0, 1.0, 0.01])..., (μ=1E-2,))
 
 include("guiding_center_3d_equations.jl")
+include("guiding_center_3d_canonical.jl")
 include("guiding_center_3d_diagnostics.jl")
 
 end
