+++ src/guiding_center_3d/guiding_center_3d_equations.jl
@@ -4,7 +4,7 @@ using Parameters
 import GeometricEquations: HODEProblem
 import GeometricSolutions: GeometricSolution, DataSeries, TimeSeries
 
-export hamiltonian
+export hamiltonian, hamiltonian_canonical
 export hode, hode_canonical
 
 
@@ -39,28 +39,6 @@ dv₂dp₃(t, q, p) = zero(eltype(p))
 dv₃dp₁(t, q, p) = zero(eltype(p))
 dv₃dp₂(t, q, p) = zero(eltype(p))
 dv₃dp₃(t, q, p) = one(eltype(p))
-
-
-d²v₁dq₁dq₁(t, q, p) = -d²A₁dx₁dx₁(t, q)
-d²v₁dq₁dq₂(t, q, p) = -d²A₁dx₁dx₂(t, q)
-d²v₁dq₁dq₃(t, q, p) = -d²A₁dx₁dx₃(t, q)
-d²v₁dq₂dq₂(t, q, p) = -d²A₁dx₂dx₂(t, q)
-d²v₁dq₂dq₃(t, q, p) = -d²A₁dx₂dx₃(t, q)
-d²v₁dq₃dq₃(t, q, p) = -d²A₁dx₃dx₃(t, q)
-
-d²v₂dq₁dq₁(t, q, p) = -d²A₂dx₁dx₁(t, q)
-d²v₂dq₁dq₂(t, q, p) = -d²A₂dx₁dx₂(t, q)
-d²v₂dq₁dq₃(t, q, p) = -d²A₂dx₁dx₃(t, q)
-d²v₂dq₂dq₂(t, q, p) = -d²A₂dx₂dx₂(t, q)
-d²v₂dq₂dq₃(t, q, p) = -d²A₂dx₂dx₃(t, q)
-d²v₂dq₃dq₃(t, q, p) = -d²A₂dx₃dx₃(t, q)
-
-d²v₃dq₁dq₁(t, q, p) = -d²A₃dx₁dx₁(t, q)
-d²v₃dq₁dq₂(t, q, p) = -d²A₃dx₁dx₂(t, q)
-d²v₃dq₁dq₃(t, q, p) = -d²A₃dx₁dx₃(t, q)
-d²v₃dq₂dq₂(t, q, p) = -d²A₃dx₂dx₂(t, q)
-d²v₃dq₂dq₃(t, q, p) = -d²A₃dx₂dx₃(t, q)
-d²v₃dq₃dq₃(t, q, p) = -d²A₃dx₃dx₃(t, q)
 
 
 u(t, q, p) = v₁(t, q, p) * g¹¹(t, q) * b₁(t, q) + v₂(t, q, p) * g²²(t, q) * b₂(t, q) + v₃(t, q, p) * g³³(t, q) * b₃(t, q)
@@ -117,6 +95,9 @@ guiding_center_3d_periodicity(::AbstractArray{T}, periodic=true) where {T<:Numbe
 
 hamiltonian(t, q, p, params) = g¹¹(t, q) * v₁(t, q, p)^2 / 2 + g²²(t, q) * v₂(t, q, p)^2 / 2 + g³³(t, q) * v₃(t, q, p)^2 / 2 + params.μ * B(t, q) + φ(t, q)
 hamiltonian_u(t, q, p, params) = u(t, q, p)^2 / 2 + params.μ * B(t, q)
+hamiltonian_canonical(t, q, p, params) = hamiltonian(t, q, p, params) + λ₁(t, q, p, params) * g₁(t, q, p) +
+                                         λ₂(t, q, p, params) * g₂(t, q, p)
+
 
 dHdq₁(t, q, p, params) = v₁(t, q, p) * g¹¹(t, q) * dv₁dq₁(t, q, p) +
                          v₂(t, q, p) * g²²(t, q) * dv₂dq₁(t, q, p) +
@@ -172,384 +153,13 @@ dHdp₃(t, q, p, params) = v₁(t, q, p) * g¹¹(t, q) * dv₁dp₃(t, q, p) +
 #     nothing
 # end
 
-
-d²Hdq₁dq₁(t, q, p, params) = (
-    dv₁dq₁(t, q, p) * g¹¹(t, q) * dv₁dq₁(t, q, p) +
-    dv₂dq₁(t, q, p) * g²²(t, q) * dv₂dq₁(t, q, p) +
-    dv₃dq₁(t, q, p) * g³³(t, q) * dv₃dq₁(t, q, p) +
-    v₁(t, q, p) * g¹¹(t, q) * d²v₁dq₁dq₁(t, q, p) +
-    v₂(t, q, p) * g²²(t, q) * d²v₂dq₁dq₁(t, q, p) +
-    v₃(t, q, p) * g³³(t, q) * d²v₃dq₁dq₁(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₁(t, q) * dv₁dq₁(t, q, p) * 2 +
-    v₂(t, q, p) * dg²²dx₁(t, q) * dv₂dq₁(t, q, p) * 2 +
-    v₃(t, q, p) * dg³³dx₁(t, q) * dv₃dq₁(t, q, p) * 2 +
-    v₁(t, q, p) * d²g¹¹dx₁dx₁(t, q) * v₁(t, q, p) / 2 +
-    v₂(t, q, p) * d²g²²dx₁dx₁(t, q) * v₂(t, q, p) / 2 +
-    v₃(t, q, p) * d²g³³dx₁dx₁(t, q) * v₃(t, q, p) / 2 +
-    params.μ * d²Bdx₁dx₁(t, q) -
-    dE₁dx₁(t, q)
-)
-
-d²Hdq₁dq₂(t, q, p, params) = (
-    dv₁dq₂(t, q, p) * g¹¹(t, q) * dv₁dq₁(t, q, p) +
-    dv₂dq₂(t, q, p) * g²²(t, q) * dv₂dq₁(t, q, p) +
-    dv₃dq₂(t, q, p) * g³³(t, q) * dv₃dq₁(t, q, p) +
-    v₁(t, q, p) * g¹¹(t, q) * d²v₁dq₁dq₂(t, q, p) +
-    v₂(t, q, p) * g²²(t, q) * d²v₂dq₁dq₂(t, q, p) +
-    v₃(t, q, p) * g³³(t, q) * d²v₃dq₁dq₂(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₁(t, q) * dv₁dq₂(t, q, p) * 2 +
-    v₂(t, q, p) * dg²²dx₁(t, q) * dv₂dq₂(t, q, p) * 2 +
-    v₃(t, q, p) * dg³³dx₁(t, q) * dv₃dq₂(t, q, p) * 2 +
-    v₁(t, q, p) * d²g¹¹dx₁dx₂(t, q) * v₁(t, q, p) / 2 +
-    v₂(t, q, p) * d²g²²dx₁dx₂(t, q) * v₂(t, q, p) / 2 +
-    v₃(t, q, p) * d²g³³dx₁dx₂(t, q) * v₃(t, q, p) / 2 +
-    params.μ * d²Bdx₁dx₂(t, q) -
-    dE₁dx₂(t, q)
-)
-
-d²Hdq₁dq₃(t, q, p, params) = (
-    dv₁dq₃(t, q, p) * g¹¹(t, q) * dv₁dq₁(t, q, p) +
-    dv₂dq₃(t, q, p) * g²²(t, q) * dv₂dq₁(t, q, p) +
-    dv₃dq₃(t, q, p) * g³³(t, q) * dv₃dq₁(t, q, p) +
-    v₁(t, q, p) * g¹¹(t, q) * d²v₁dq₁dq₃(t, q, p) +
-    v₂(t, q, p) * g²²(t, q) * d²v₂dq₁dq₃(t, q, p) +
-    v₃(t, q, p) * g³³(t, q) * d²v₃dq₁dq₃(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₁(t, q) * dv₁dq₃(t, q, p) * 2 +
-    v₂(t, q, p) * dg²²dx₁(t, q) * dv₂dq₃(t, q, p) * 2 +
-    v₃(t, q, p) * dg³³dx₁(t, q) * dv₃dq₃(t, q, p) * 2 +
-    v₁(t, q, p) * d²g¹¹dx₁dx₃(t, q) * v₁(t, q, p) / 2 +
-    v₂(t, q, p) * d²g²²dx₁dx₃(t, q) * v₂(t, q, p) / 2 +
-    v₃(t, q, p) * d²g³³dx₁dx₃(t, q) * v₃(t, q, p) / 2 +
-    params.μ * d²Bdx₁dx₃(t, q) -
-    dE₁dx₃(t, q)
-)
-
-d²Hdq₁dp₁(t, q, p, params) = (
-    dv₁dp₁(t, q, p) * g¹¹(t, q) * dv₁dq₁(t, q, p) +
-    dv₂dp₁(t, q, p) * g²²(t, q) * dv₂dq₁(t, q, p) +
-    dv₃dp₁(t, q, p) * g³³(t, q) * dv₃dq₁(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₁(t, q) * dv₁dp₁(t, q, p) +
-    v₂(t, q, p) * dg²²dx₁(t, q) * dv₂dp₁(t, q, p) +
-    v₃(t, q, p) * dg³³dx₁(t, q) * dv₃dp₁(t, q, p)
-)
-
-d²Hdq₁dp₂(t, q, p, params) = (
-    dv₁dp₂(t, q, p) * g¹¹(t, q) * dv₁dq₁(t, q, p) +
-    dv₂dp₂(t, q, p) * g²²(t, q) * dv₂dq₁(t, q, p) +
-    dv₃dp₂(t, q, p) * g³³(t, q) * dv₃dq₁(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₁(t, q) * dv₁dp₂(t, q, p) +
-    v₂(t, q, p) * dg²²dx₁(t, q) * dv₂dp₂(t, q, p) +
-    v₃(t, q, p) * dg³³dx₁(t, q) * dv₃dp₂(t, q, p)
-)
-
-d²Hdq₁dp₃(t, q, p, params) = (
-    dv₁dp₃(t, q, p) * g¹¹(t, q) * dv₁dq₁(t, q, p) +
-    dv₂dp₃(t, q, p) * g²²(t, q) * dv₂dq₁(t, q, p) +
-    dv₃dp₃(t, q, p) * g³³(t, q) * dv₃dq₁(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₁(t, q) * dv₁dp₃(t, q, p) +
-    v₂(t, q, p) * dg²²dx₁(t, q) * dv₂dp₃(t, q, p) +
-    v₃(t, q, p) * dg³³dx₁(t, q) * dv₃dp₃(t, q, p)
-)
-
-
-d²Hdq₂dq₁(t, q, p, params) = (
-    dv₁dq₁(t, q, p) * g¹¹(t, q) * dv₁dq₂(t, q, p) +
-    dv₂dq₁(t, q, p) * g²²(t, q) * dv₂dq₂(t, q, p) +
-    dv₃dq₁(t, q, p) * g³³(t, q) * dv₃dq₂(t, q, p) +
-    v₁(t, q, p) * g¹¹(t, q) * d²v₁dq₂dq₁(t, q, p) +
-    v₂(t, q, p) * g²²(t, q) * d²v₂dq₂dq₁(t, q, p) +
-    v₃(t, q, p) * g³³(t, q) * d²v₃dq₂dq₁(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₂(t, q) * dv₁dq₁(t, q, p) * 2 +
-    v₂(t, q, p) * dg²²dx₂(t, q) * dv₂dq₁(t, q, p) * 2 +
-    v₃(t, q, p) * dg³³dx₂(t, q) * dv₃dq₁(t, q, p) * 2 +
-    v₁(t, q, p) * d²g¹¹dx₂dx₁(t, q) * v₁(t, q, p) / 2 +
-    v₂(t, q, p) * d²g²²dx₂dx₁(t, q) * v₂(t, q, p) / 2 +
-    v₃(t, q, p) * d²g³³dx₂dx₁(t, q) * v₃(t, q, p) / 2 +
-    params.μ * d²Bdx₂dx₁(t, q) -
-    dE₂dx₁(t, q)
-)
-
-d²Hdq₂dq₂(t, q, p, params) = (
-    dv₁dq₂(t, q, p) * g¹¹(t, q) * dv₁dq₂(t, q, p) +
-    dv₂dq₂(t, q, p) * g²²(t, q) * dv₂dq₂(t, q, p) +
-    dv₃dq₂(t, q, p) * g³³(t, q) * dv₃dq₂(t, q, p) +
-    v₁(t, q, p) * g¹¹(t, q) * d²v₁dq₂dq₂(t, q, p) +
-    v₂(t, q, p) * g²²(t, q) * d²v₂dq₂dq₂(t, q, p) +
-    v₃(t, q, p) * g³³(t, q) * d²v₃dq₂dq₂(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₂(t, q) * dv₁dq₂(t, q, p) * 2 +
-    v₂(t, q, p) * dg²²dx₂(t, q) * dv₂dq₂(t, q, p) * 2 +
-    v₃(t, q, p) * dg³³dx₂(t, q) * dv₃dq₂(t, q, p) * 2 +
-    v₁(t, q, p) * d²g¹¹dx₂dx₂(t, q) * v₁(t, q, p) / 2 +
-    v₂(t, q, p) * d²g²²dx₂dx₂(t, q) * v₂(t, q, p) / 2 +
-    v₃(t, q, p) * d²g³³dx₂dx₂(t, q) * v₃(t, q, p) / 2 +
-    params.μ * d²Bdx₂dx₂(t, q) -
-    dE₂dx₂(t, q)
-)
-
-d²Hdq₂dq₃(t, q, p, params) = (
-    dv₁dq₃(t, q, p) * g¹¹(t, q) * dv₁dq₂(t, q, p) +
-    dv₂dq₃(t, q, p) * g²²(t, q) * dv₂dq₂(t, q, p) +
-    dv₃dq₃(t, q, p) * g³³(t, q) * dv₃dq₂(t, q, p) +
-    v₁(t, q, p) * g¹¹(t, q) * d²v₁dq₂dq₃(t, q, p) +
-    v₂(t, q, p) * g²²(t, q) * d²v₂dq₂dq₃(t, q, p) +
-    v₃(t, q, p) * g³³(t, q) * d²v₃dq₂dq₃(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₂(t, q) * dv₁dq₃(t, q, p) * 2 +
-    v₂(t, q, p) * dg²²dx₂(t, q) * dv₂dq₃(t, q, p) * 2 +
-    v₃(t, q, p) * dg³³dx₂(t, q) * dv₃dq₃(t, q, p) * 2 +
-    v₁(t, q, p) * d²g¹¹dx₂dx₃(t, q) * v₁(t, q, p) / 2 +
-    v₂(t, q, p) * d²g²²dx₂dx₃(t, q) * v₂(t, q, p) / 2 +
-    v₃(t, q, p) * d²g³³dx₂dx₃(t, q) * v₃(t, q, p) / 2 +
-    params.μ * d²Bdx₂dx₃(t, q) -
-    dE₂dx₃(t, q)
-)
-
-d²Hdq₂dp₁(t, q, p, params) = (
-    dv₁dp₁(t, q, p) * g¹¹(t, q) * dv₁dq₂(t, q, p) +
-    dv₂dp₁(t, q, p) * g²²(t, q) * dv₂dq₂(t, q, p) +
-    dv₃dp₁(t, q, p) * g³³(t, q) * dv₃dq₂(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₂(t, q) * dv₁dp₁(t, q, p) +
-    v₂(t, q, p) * dg²²dx₂(t, q) * dv₂dp₁(t, q, p) +
-    v₃(t, q, p) * dg³³dx₂(t, q) * dv₃dp₁(t, q, p)
-)
-
-d²Hdq₂dp₂(t, q, p, params) = (
-    dv₁dp₂(t, q, p) * g¹¹(t, q) * dv₁dq₂(t, q, p) +
-    dv₂dp₂(t, q, p) * g²²(t, q) * dv₂dq₂(t, q, p) +
-    dv₃dp₂(t, q, p) * g³³(t, q) * dv₃dq₂(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₂(t, q) * dv₁dp₂(t, q, p) +
-    v₂(t, q, p) * dg²²dx₂(t, q) * dv₂dp₂(t, q, p) +
-    v₃(t, q, p) * dg³³dx₂(t, q) * dv₃dp₂(t, q, p)
-)
-
-d²Hdq₂dp₃(t, q, p, params) = (
-    dv₁dp₃(t, q, p) * g¹¹(t, q) * dv₁dq₂(t, q, p) +
-    dv₂dp₃(t, q, p) * g²²(t, q) * dv₂dq₂(t, q, p) +
-    dv₃dp₃(t, q, p) * g³³(t, q) * dv₃dq₂(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₂(t, q) * dv₁dp₃(t, q, p) +
-    v₂(t, q, p) * dg²²dx₂(t, q) * dv₂dp₃(t, q, p) +
-    v₃(t, q, p) * dg³³dx₂(t, q) * dv₃dp₃(t, q, p)
-)
-
-
-d²Hdq₃dq₁(t, q, p, params) = (
-    dv₁dq₁(t, q, p) * g¹¹(t, q) * dv₁dq₃(t, q, p) +
-    dv₂dq₁(t, q, p) * g²²(t, q) * dv₂dq₃(t, q, p) +
-    dv₃dq₁(t, q, p) * g³³(t, q) * dv₃dq₃(t, q, p) +
-    v₁(t, q, p) * g¹¹(t, q) * d²v₁dq₃dq₁(t, q, p) +
-    v₂(t, q, p) * g²²(t, q) * d²v₂dq₃dq₁(t, q, p) +
-    v₃(t, q, p) * g³³(t, q) * d²v₃dq₃dq₁(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₃(t, q) * dv₁dq₁(t, q, p) * 2 +
-    v₂(t, q, p) * dg²²dx₃(t, q) * dv₂dq₁(t, q, p) * 2 +
-    v₃(t, q, p) * dg³³dx₃(t, q) * dv₃dq₁(t, q, p) * 2 +
-    v₁(t, q, p) * d²g¹¹dx₃dx₁(t, q) * v₁(t, q, p) / 2 +
-    v₂(t, q, p) * d²g²²dx₃dx₁(t, q) * v₂(t, q, p) / 2 +
-    v₃(t, q, p) * d²g³³dx₃dx₁(t, q) * v₃(t, q, p) / 2 +
-    params.μ * d²Bdx₃dx₁(t, q) -
-    dE₃dx₁(t, q)
-)
-
-d²Hdq₃dq₂(t, q, p, params) = (
-    dv₁dq₂(t, q, p) * g¹¹(t, q) * dv₁dq₃(t, q, p) +
-    dv₂dq₂(t, q, p) * g²²(t, q) * dv₂dq₃(t, q, p) +
-    dv₃dq₂(t, q, p) * g³³(t, q) * dv₃dq₃(t, q, p) +
-    v₁(t, q, p) * g¹¹(t, q) * d²v₁dq₃dq₂(t, q, p) +
-    v₂(t, q, p) * g²²(t, q) * d²v₂dq₃dq₂(t, q, p) +
-    v₃(t, q, p) * g³³(t, q) * d²v₃dq₃dq₂(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₃(t, q) * dv₁dq₂(t, q, p) * 2 +
-    v₂(t, q, p) * dg²²dx₃(t, q) * dv₂dq₂(t, q, p) * 2 +
-    v₃(t, q, p) * dg³³dx₃(t, q) * dv₃dq₂(t, q, p) * 2 +
-    v₁(t, q, p) * d²g¹¹dx₃dx₂(t, q) * v₁(t, q, p) / 2 +
-    v₂(t, q, p) * d²g²²dx₃dx₂(t, q) * v₂(t, q, p) / 2 +
-    v₃(t, q, p) * d²g³³dx₃dx₂(t, q) * v₃(t, q, p) / 2 +
-    params.μ * d²Bdx₃dx₂(t, q) -
-    dE₃dx₂(t, q)
-)
-
-d²Hdq₃dq₃(t, q, p, params) = (
-    dv₁dq₃(t, q, p) * g¹¹(t, q) * dv₁dq₃(t, q, p) +
-    dv₂dq₃(t, q, p) * g²²(t, q) * dv₂dq₃(t, q, p) +
-    dv₃dq₃(t, q, p) * g³³(t, q) * dv₃dq₃(t, q, p) +
-    v₁(t, q, p) * g¹¹(t, q) * d²v₁dq₃dq₃(t, q, p) +
-    v₂(t, q, p) * g²²(t, q) * d²v₂dq₃dq₃(t, q, p) +
-    v₃(t, q, p) * g³³(t, q) * d²v₃dq₃dq₃(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₃(t, q) * dv₁dq₃(t, q, p) * 2 +
-    v₂(t, q, p) * dg²²dx₃(t, q) * dv₂dq₃(t, q, p) * 2 +
-    v₃(t, q, p) * dg³³dx₃(t, q) * dv₃dq₃(t, q, p) * 2 +
-    v₁(t, q, p) * d²g¹¹dx₃dx₃(t, q) * v₁(t, q, p) / 2 +
-    v₂(t, q, p) * d²g²²dx₃dx₃(t, q) * v₂(t, q, p) / 2 +
-    v₃(t, q, p) * d²g³³dx₃dx₃(t, q) * v₃(t, q, p) / 2 +
-    params.μ * d²Bdx₃dx₃(t, q) -
-    dE₃dx₃(t, q)
-)
-
-d²Hdq₃dp₁(t, q, p, params) = (
-    dv₁dp₁(t, q, p) * g¹¹(t, q) * dv₁dq₃(t, q, p) +
-    dv₂dp₁(t, q, p) * g²²(t, q) * dv₂dq₃(t, q, p) +
-    dv₃dp₁(t, q, p) * g³³(t, q) * dv₃dq₃(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₃(t, q) * dv₁dp₁(t, q, p) +
-    v₂(t, q, p) * dg²²dx₃(t, q) * dv₂dp₁(t, q, p) +
-    v₃(t, q, p) * dg³³dx₃(t, q) * dv₃dp₁(t, q, p)
-)
-
-d²Hdq₃dp₂(t, q, p, params) = (
-    dv₁dp₂(t, q, p) * g¹¹(t, q) * dv₁dq₃(t, q, p) +
-    dv₂dp₂(t, q, p) * g²²(t, q) * dv₂dq₃(t, q, p) +
-    dv₃dp₂(t, q, p) * g³³(t, q) * dv₃dq₃(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₃(t, q) * dv₁dp₂(t, q, p) +
-    v₂(t, q, p) * dg²²dx₃(t, q) * dv₂dp₂(t, q, p) +
-    v₃(t, q, p) * dg³³dx₃(t, q) * dv₃dp₂(t, q, p)
-)
-
-d²Hdq₃dp₃(t, q, p, params) = (
-    dv₁dp₃(t, q, p) * g¹¹(t, q) * dv₁dq₃(t, q, p) +
-    dv₂dp₃(t, q, p) * g²²(t, q) * dv₂dq₃(t, q, p) +
-    dv₃dp₃(t, q, p) * g³³(t, q) * dv₃dq₃(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₃(t, q) * dv₁dp₃(t, q, p) +
-    v₂(t, q, p) * dg²²dx₃(t, q) * dv₂dp₃(t, q, p) +
-    v₃(t, q, p) * dg³³dx₃(t, q) * dv₃dp₃(t, q, p)
-)
-
-
-d²Hdp₁dq₁(t, q, p, params) = (
-    dv₁dq₁(t, q, p) * g¹¹(t, q) * dv₁dp₁(t, q, p) +
-    dv₂dq₁(t, q, p) * g²²(t, q) * dv₂dp₁(t, q, p) +
-    dv₃dq₁(t, q, p) * g³³(t, q) * dv₃dp₁(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₁(t, q) * dv₁dp₁(t, q, p) +
-    v₂(t, q, p) * dg²²dx₁(t, q) * dv₂dp₁(t, q, p) +
-    v₃(t, q, p) * dg³³dx₁(t, q) * dv₃dp₁(t, q, p)
-)
-
-d²Hdp₂dq₁(t, q, p, params) = (
-    dv₁dq₁(t, q, p) * g¹¹(t, q) * dv₁dp₂(t, q, p) +
-    dv₂dq₁(t, q, p) * g²²(t, q) * dv₂dp₂(t, q, p) +
-    dv₃dq₁(t, q, p) * g³³(t, q) * dv₃dp₂(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₁(t, q) * dv₁dp₂(t, q, p) +
-    v₂(t, q, p) * dg²²dx₁(t, q) * dv₂dp₂(t, q, p) +
-    v₃(t, q, p) * dg³³dx₁(t, q) * dv₃dp₂(t, q, p)
-)
-
-d²Hdp₃dq₁(t, q, p, params) = (
-    dv₁dq₁(t, q, p) * g¹¹(t, q) * dv₁dp₃(t, q, p) +
-    dv₂dq₁(t, q, p) * g²²(t, q) * dv₂dp₃(t, q, p) +
-    dv₃dq₁(t, q, p) * g³³(t, q) * dv₃dp₃(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₁(t, q) * dv₁dp₃(t, q, p) +
-    v₂(t, q, p) * dg²²dx₁(t, q) * dv₂dp₃(t, q, p) +
-    v₃(t, q, p) * dg³³dx₁(t, q) * dv₃dp₃(t, q, p)
-)
-
-d²Hdp₁dp₁(t, q, p, params) = (
-    dv₁dp₁(t, q, p) * g¹¹(t, q) * dv₁dp₁(t, q, p) +
-    dv₂dp₁(t, q, p) * g²²(t, q) * dv₂dp₁(t, q, p) +
-    dv₃dp₁(t, q, p) * g³³(t, q) * dv₃dp₁(t, q, p)
-)
-
-d²Hdp₂dp₁(t, q, p, params) = (
-    dv₁dp₁(t, q, p) * g¹¹(t, q) * dv₁dp₂(t, q, p) +
-    dv₂dp₁(t, q, p) * g²²(t, q) * dv₂dp₂(t, q, p) +
-    dv₃dp₁(t, q, p) * g³³(t, q) * dv₃dp₂(t, q, p)
-)
-
-d²Hdp₃dp₁(t, q, p, params) = (
-    dv₁dp₁(t, q, p) * g¹¹(t, q) * dv₁dp₃(t, q, p) +
-    dv₂dp₁(t, q, p) * g²²(t, q) * dv₂dp₃(t, q, p) +
-    dv₃dp₁(t, q, p) * g³³(t, q) * dv₃dp₃(t, q, p)
-)
-
-
-d²Hdp₁dq₂(t, q, p, params) = (
-    dv₁dq₂(t, q, p) * g¹¹(t, q) * dv₁dp₁(t, q, p) +
-    dv₂dq₂(t, q, p) * g²²(t, q) * dv₂dp₁(t, q, p) +
-    dv₃dq₂(t, q, p) * g³³(t, q) * dv₃dp₁(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₂(t, q) * dv₁dp₁(t, q, p) +
-    v₂(t, q, p) * dg²²dx₂(t, q) * dv₂dp₁(t, q, p) +
-    v₃(t, q, p) * dg³³dx₂(t, q) * dv₃dp₁(t, q, p)
-)
-
-d²Hdp₂dq₂(t, q, p, params) = (
-    dv₁dq₂(t, q, p) * g¹¹(t, q) * dv₁dp₂(t, q, p) +
-    dv₂dq₂(t, q, p) * g²²(t, q) * dv₂dp₂(t, q, p) +
-    dv₃dq₂(t, q, p) * g³³(t, q) * dv₃dp₂(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₂(t, q) * dv₁dp₂(t, q, p) +
-    v₂(t, q, p) * dg²²dx₂(t, q) * dv₂dp₂(t, q, p) +
-    v₃(t, q, p) * dg³³dx₂(t, q) * dv₃dp₂(t, q, p)
-)
-
-d²Hdp₃dq₂(t, q, p, params) = (
-    dv₁dq₂(t, q, p) * g¹¹(t, q) * dv₁dp₃(t, q, p) +
-    dv₂dq₂(t, q, p) * g²²(t, q) * dv₂dp₃(t, q, p) +
-    dv₃dq₂(t, q, p) * g³³(t, q) * dv₃dp₃(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₂(t, q) * dv₁dp₃(t, q, p) +
-    v₂(t, q, p) * dg²²dx₂(t, q) * dv₂dp₃(t, q, p) +
-    v₃(t, q, p) * dg³³dx₂(t, q) * dv₃dp₃(t, q, p)
-)
-
-d²Hdp₁dp₂(t, q, p, params) = (
-    dv₁dp₂(t, q, p) * g¹¹(t, q) * dv₁dp₁(t, q, p) +
-    dv₂dp₂(t, q, p) * g²²(t, q) * dv₂dp₁(t, q, p) +
-    dv₃dp₂(t, q, p) * g³³(t, q) * dv₃dp₁(t, q, p)
-)
-
-d²Hdp₂dp₂(t, q, p, params) = (
-    dv₁dp₂(t, q, p) * g¹¹(t, q) * dv₁dp₂(t, q, p) +
-    dv₂dp₂(t, q, p) * g²²(t, q) * dv₂dp₂(t, q, p) +
-    dv₃dp₂(t, q, p) * g³³(t, q) * dv₃dp₂(t, q, p)
-)
-
-d²Hdp₃dp₂(t, q, p, params) = (
-    dv₁dp₂(t, q, p) * g¹¹(t, q) * dv₁dp₃(t, q, p) +
-    dv₂dp₂(t, q, p) * g²²(t, q) * dv₂dp₃(t, q, p) +
-    dv₃dp₂(t, q, p) * g³³(t, q) * dv₃dp₃(t, q, p)
-)
-
-
-d²Hdp₁dq₃(t, q, p, params) = (
-    dv₁dq₃(t, q, p) * g¹¹(t, q) * dv₁dp₁(t, q, p) +
-    dv₂dq₃(t, q, p) * g²²(t, q) * dv₂dp₁(t, q, p) +
-    dv₃dq₃(t, q, p) * g³³(t, q) * dv₃dp₁(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₃(t, q) * dv₁dp₁(t, q, p) +
-    v₂(t, q, p) * dg²²dx₃(t, q) * dv₂dp₁(t, q, p) +
-    v₃(t, q, p) * dg³³dx₃(t, q) * dv₃dp₁(t, q, p)
-)
-
-d²Hdp₂dq₃(t, q, p, params) = (
-    dv₁dq₃(t, q, p) * g¹¹(t, q) * dv₁dp₂(t, q, p) +
-    dv₂dq₃(t, q, p) * g²²(t, q) * dv₂dp₂(t, q, p) +
-    dv₃dq₃(t, q, p) * g³³(t, q) * dv₃dp₂(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₃(t, q) * dv₁dp₂(t, q, p) +
-    v₂(t, q, p) * dg²²dx₃(t, q) * dv₂dp₂(t, q, p) +
-    v₃(t, q, p) * dg³³dx₃(t, q) * dv₃dp₂(t, q, p)
-)
-
-d²Hdp₃dq₃(t, q, p, params) = (
-    dv₁dq₃(t, q, p) * g¹¹(t, q) * dv₁dp₃(t, q, p) +
-    dv₂dq₃(t, q, p) * g²²(t, q) * dv₂dp₃(t, q, p) +
-    dv₃dq₃(t, q, p) * g³³(t, q) * dv₃dp₃(t, q, p) +
-    v₁(t, q, p) * dg¹¹dx₃(t, q) * dv₁dp₃(t, q, p) +
-    v₂(t, q, p) * dg²²dx₃(t, q) * dv₂dp₃(t, q, p) +
-    v₃(t, q, p) * dg³³dx₃(t, q) * dv₃dp₃(t, q, p)
-)
-
-d²Hdp₁dp₃(t, q, p, params) = (
-    dv₁dp₃(t, q, p) * g¹¹(t, q) * dv₁dp₁(t, q, p) +
-    dv₂dp₃(t, q, p) * g²²(t, q) * dv₂dp₁(t, q, p) +
-    dv₃dp₃(t, q, p) * g³³(t, q) * dv₃dp₁(t, q, p)
-)
-
-d²Hdp₂dp₃(t, q, p, params) = (
-    dv₁dp₃(t, q, p) * g¹¹(t, q) * dv₁dp₂(t, q, p) +
-    dv₂dp₃(t, q, p) * g²²(t, q) * dv₂dp₂(t, q, p) +
-    dv₃dp₃(t, q, p) * g³³(t, q) * dv₃dp₂(t, q, p)
-)
-
-d²Hdp₃dp₃(t, q, p, params) = (
-    dv₁dp₃(t, q, p) * g¹¹(t, q) * dv₁dp₃(t, q, p) +
-    dv₂dp₃(t, q, p) * g²²(t, q) * dv₂dp₃(t, q, p) +
-    dv₃dp₃(t, q, p) * g³³(t, q) * dv₃dp₃(t, q, p)
-)
-
-
 g₁(t, q, p) = b₁(t, q) * v₂(t, q, p) - b₂(t, q) * v₁(t, q, p)
 g₂(t, q, p) = b₁(t, q) * v₃(t, q, p) - b₃(t, q) * v₁(t, q, p)
+# g₁(t, q, p) = b₁(t, q) * v₃(t, q, p) - b₃(t, q) * v₁(t, q, p)
+# g₂(t, q, p) = b₂(t, q) * v₃(t, q, p) - b₃(t, q) * v₂(t, q, p)
+
+g₁(t, q, p, params) = g₁(t, q, p)
+g₂(t, q, p, params) = g₂(t, q, p)
 
 dg₁dq₁(t, q, p) = (db₁dx₁(t, q) * v₂(t, q, p) + b₁(t, q) * dv₂dq₁(t, q, p)) -
                   (db₂dx₁(t, q) * v₁(t, q, p) + b₂(t, q) * dv₁dq₁(t, q, p))
@@ -574,153 +184,27 @@ dg₂dp₂(t, q, p) = b₁(t, q) * dv₃dp₂(t, q, p) - b₃(t, q) * dv₁dp₂
 dg₂dp₃(t, q, p) = b₁(t, q) * dv₃dp₃(t, q, p) - b₃(t, q) * dv₁dp₃(t, q, p)
 
 
+# dg₁dq₁(t, q, p) = (db₁dx₁(t, q) * v₃(t, q, p) + b₁(t, q) * dv₃dq₁(t, q, p)) -
+#                   (db₃dx₁(t, q) * v₁(t, q, p) + b₃(t, q) * dv₁dq₁(t, q, p))
+# dg₁dq₂(t, q, p) = (db₁dx₂(t, q) * v₃(t, q, p) + b₁(t, q) * dv₃dq₂(t, q, p)) -
+#                   (db₃dx₂(t, q) * v₁(t, q, p) + b₃(t, q) * dv₁dq₂(t, q, p))
+# dg₁dq₃(t, q, p) = (db₁dx₃(t, q) * v₃(t, q, p) + b₁(t, q) * dv₃dq₃(t, q, p)) -
+#                   (db₃dx₃(t, q) * v₁(t, q, p) + b₃(t, q) * dv₁dq₃(t, q, p))
 
-d²g₁dq₁dq₁(t, q, p) = (d²b₁dx₁dx₁(t, q) * v₂(t, q, p) + b₁(t, q) * d²v₂dq₁dq₁(t, q, p)) +
-                      (d²b₁dx₁dx₁(t, q) * v₂(t, q, p) + b₁(t, q) * d²v₂dq₁dq₁(t, q, p)) -
-                      (db₂dx₁(t, q) * dv₁dq₁(t, q, p) + db₂dx₁(t, q) * dv₁dq₁(t, q, p)) -
-                      (db₂dx₁(t, q) * dv₁dq₁(t, q, p) + db₂dx₁(t, q) * dv₁dq₁(t, q, p))
+# dg₂dq₁(t, q, p) = (db₂dx₁(t, q) * v₃(t, q, p) + b₂(t, q) * dv₃dq₁(t, q, p)) -
+#                   (db₃dx₁(t, q) * v₂(t, q, p) + b₃(t, q) * dv₂dq₁(t, q, p))
+# dg₂dq₂(t, q, p) = (db₂dx₂(t, q) * v₃(t, q, p) + b₂(t, q) * dv₃dq₂(t, q, p)) -
+#                   (db₃dx₂(t, q) * v₂(t, q, p) + b₃(t, q) * dv₂dq₂(t, q, p))
+# dg₂dq₃(t, q, p) = (db₂dx₃(t, q) * v₃(t, q, p) + b₂(t, q) * dv₃dq₃(t, q, p)) -
+#                   (db₃dx₃(t, q) * v₂(t, q, p) + b₃(t, q) * dv₂dq₃(t, q, p))
 
-d²g₁dq₁dq₂(t, q, p) = (d²b₁dx₁dx₂(t, q) * v₂(t, q, p) + b₁(t, q) * d²v₂dq₁dq₂(t, q, p)) +
-                      (d²b₁dx₁dx₂(t, q) * v₂(t, q, p) + b₁(t, q) * d²v₂dq₁dq₂(t, q, p)) -
-                      (db₂dx₂(t, q) * dv₁dq₁(t, q, p) + db₂dx₁(t, q) * dv₁dq₂(t, q, p)) -
-                      (db₂dx₂(t, q) * dv₁dq₁(t, q, p) + db₂dx₁(t, q) * dv₁dq₂(t, q, p))
+# dg₁dp₁(t, q, p) = b₁(t, q) * dv₃dp₁(t, q, p) - b₃(t, q) * dv₁dp₁(t, q, p)
+# dg₁dp₂(t, q, p) = b₁(t, q) * dv₃dp₂(t, q, p) - b₃(t, q) * dv₁dp₂(t, q, p)
+# dg₁dp₃(t, q, p) = b₁(t, q) * dv₃dp₃(t, q, p) - b₃(t, q) * dv₁dp₃(t, q, p)
 
-d²g₁dq₁dq₃(t, q, p) = (d²b₁dx₁dx₃(t, q) * v₂(t, q, p) + b₁(t, q) * d²v₂dq₁dq₃(t, q, p)) +
-                      (d²b₁dx₁dx₃(t, q) * v₂(t, q, p) + b₁(t, q) * d²v₂dq₁dq₃(t, q, p)) -
-                      (db₂dx₃(t, q) * dv₁dq₁(t, q, p) + db₂dx₁(t, q) * dv₁dq₃(t, q, p)) -
-                      (db₂dx₃(t, q) * dv₁dq₁(t, q, p) + db₂dx₁(t, q) * dv₁dq₃(t, q, p))
-
-
-d²g₁dq₂dq₁(t, q, p) = (d²b₁dx₁dx₂(t, q) * v₂(t, q, p) + b₁(t, q) * d²v₂dq₁dq₂(t, q, p)) +
-                      (d²b₁dx₁dx₂(t, q) * v₂(t, q, p) + b₁(t, q) * d²v₂dq₁dq₂(t, q, p)) -
-                      (db₂dx₁(t, q) * dv₁dq₂(t, q, p) + db₂dx₂(t, q) * dv₁dq₁(t, q, p)) -
-                      (db₂dx₁(t, q) * dv₁dq₂(t, q, p) + db₂dx₂(t, q) * dv₁dq₁(t, q, p))
-
-d²g₁dq₂dq₂(t, q, p) = (d²b₁dx₁dx₂(t, q) * v₂(t, q, p) + b₁(t, q) * d²v₂dq₁dq₂(t, q, p)) +
-                      (d²b₁dx₁dx₂(t, q) * v₂(t, q, p) + b₁(t, q) * d²v₂dq₁dq₂(t, q, p)) -
-                      (db₂dx₂(t, q) * dv₁dq₂(t, q, p) + db₂dx₂(t, q) * dv₁dq₂(t, q, p)) -
-                      (db₂dx₂(t, q) * dv₁dq₂(t, q, p) + db₂dx₂(t, q) * dv₁dq₂(t, q, p))
-
-d²g₁dq₂dq₃(t, q, p) = (d²b₁dx₁dx₃(t, q) * v₂(t, q, p) + b₁(t, q) * d²v₂dq₁dq₃(t, q, p)) +
-                      (d²b₁dx₁dx₃(t, q) * v₂(t, q, p) + b₁(t, q) * d²v₂dq₁dq₃(t, q, p)) -
-                      (db₂dx₃(t, q) * dv₁dq₂(t, q, p) + db₂dx₂(t, q) * dv₁dq₃(t, q, p)) -
-                      (db₂dx₃(t, q) * dv₁dq₂(t, q, p) + db₂dx₂(t, q) * dv₁dq₃(t, q, p))
-
-
-d²g₁dq₃dq₁(t, q, p) = (d²b₁dx₁dx₃(t, q) * v₂(t, q, p) + b₁(t, q) * d²v₂dq₁dq₃(t, q, p)) +
-                      (d²b₁dx₁dx₃(t, q) * v₂(t, q, p) + b₁(t, q) * d²v₂dq₁dq₃(t, q, p)) -
-                      (db₂dx₁(t, q) * dv₁dq₃(t, q, p) + db₂dx₃(t, q) * dv₁dq₁(t, q, p)) -
-                      (db₂dx₁(t, q) * dv₁dq₃(t, q, p) + db₂dx₃(t, q) * dv₁dq₁(t, q, p))
-
-d²g₁dq₃dq₂(t, q, p) = (d²b₁dx₂dx₃(t, q) * v₂(t, q, p) + b₁(t, q) * d²v₂dq₂dq₃(t, q, p)) +
-                      (d²b₁dx₂dx₃(t, q) * v₂(t, q, p) + b₁(t, q) * d²v₂dq₂dq₃(t, q, p)) -
-                      (db₂dx₂(t, q) * dv₁dq₃(t, q, p) + db₂dx₃(t, q) * dv₁dq₂(t, q, p)) -
-                      (db₂dx₂(t, q) * dv₁dq₃(t, q, p) + db₂dx₃(t, q) * dv₁dq₂(t, q, p))
-
-d²g₁dq₃dq₃(t, q, p) = (d²b₁dx₃dx₃(t, q) * v₂(t, q, p) + b₁(t, q) * d²v₂dq₃dq₃(t, q, p)) +
-                      (d²b₁dx₃dx₃(t, q) * v₂(t, q, p) + b₁(t, q) * d²v₂dq₃dq₃(t, q, p)) -
-                      (db₂dx₃(t, q) * dv₁dq₃(t, q, p) + db₂dx₃(t, q) * dv₁dq₃(t, q, p)) -
-                      (db₂dx₃(t, q) * dv₁dq₃(t, q, p) + db₂dx₃(t, q) * dv₁dq₃(t, q, p))
-
-
-d²g₂dq₁dq₁(t, q, p) = (d²b₁dx₁dx₁(t, q) * v₃(t, q, p) + b₁(t, q) * d²v₃dq₁dq₁(t, q, p)) +
-                      (d²b₁dx₁dx₁(t, q) * v₃(t, q, p) + b₁(t, q) * d²v₃dq₁dq₁(t, q, p)) -
-                      (db₃dx₁(t, q) * dv₁dq₁(t, q, p) + db₃dx₁(t, q) * dv₁dq₁(t, q, p)) -
-                      (db₃dx₁(t, q) * dv₁dq₁(t, q, p) + db₃dx₁(t, q) * dv₁dq₁(t, q, p))
-
-d²g₂dq₁dq₂(t, q, p) = (d²b₁dx₁dx₂(t, q) * v₃(t, q, p) + b₁(t, q) * d²v₃dq₁dq₂(t, q, p)) +
-                      (d²b₁dx₁dx₂(t, q) * v₃(t, q, p) + b₁(t, q) * d²v₃dq₁dq₂(t, q, p)) -
-                      (db₃dx₂(t, q) * dv₁dq₁(t, q, p) + db₃dx₁(t, q) * dv₁dq₂(t, q, p)) -
-                      (db₃dx₂(t, q) * dv₁dq₁(t, q, p) + db₃dx₁(t, q) * dv₁dq₂(t, q, p))
-
-d²g₂dq₁dq₃(t, q, p) = (d²b₁dx₁dx₃(t, q) * v₃(t, q, p) + b₁(t, q) * d²v₃dq₁dq₃(t, q, p)) +
-                      (d²b₁dx₁dx₃(t, q) * v₃(t, q, p) + b₁(t, q) * d²v₃dq₁dq₃(t, q, p)) -
-                      (db₃dx₃(t, q) * dv₁dq₁(t, q, p) + db₃dx₁(t, q) * dv₁dq₃(t, q, p)) -
-                      (db₃dx₃(t, q) * dv₁dq₁(t, q, p) + db₃dx₁(t, q) * dv₁dq₃(t, q, p))
-
-
-d²g₂dq₂dq₁(t, q, p) = (d²b₁dx₁dx₂(t, q) * v₃(t, q, p) + b₁(t, q) * d²v₃dq₁dq₂(t, q, p)) +
-                      (d²b₁dx₁dx₂(t, q) * v₃(t, q, p) + b₁(t, q) * d²v₃dq₁dq₂(t, q, p)) -
-                      (db₃dx₁(t, q) * dv₁dq₂(t, q, p) + db₃dx₂(t, q) * dv₁dq₁(t, q, p)) -
-                      (db₃dx₁(t, q) * dv₁dq₂(t, q, p) + db₃dx₂(t, q) * dv₁dq₁(t, q, p))
-
-d²g₂dq₂dq₂(t, q, p) = (d²b₁dx₂dx₂(t, q) * v₃(t, q, p) + b₁(t, q) * d²v₃dq₂dq₂(t, q, p)) +
-                      (d²b₁dx₂dx₂(t, q) * v₃(t, q, p) + b₁(t, q) * d²v₃dq₂dq₂(t, q, p)) -
-                      (db₃dx₂(t, q) * dv₁dq₂(t, q, p) + db₃dx₂(t, q) * dv₁dq₂(t, q, p)) -
-                      (db₃dx₂(t, q) * dv₁dq₂(t, q, p) + db₃dx₂(t, q) * dv₁dq₂(t, q, p))
-
-d²g₂dq₂dq₃(t, q, p) = (d²b₁dx₂dx₃(t, q) * v₃(t, q, p) + b₁(t, q) * d²v₃dq₂dq₃(t, q, p)) +
-                      (d²b₁dx₂dx₃(t, q) * v₃(t, q, p) + b₁(t, q) * d²v₃dq₂dq₃(t, q, p)) -
-                      (db₃dx₃(t, q) * dv₁dq₂(t, q, p) + db₃dx₂(t, q) * dv₁dq₃(t, q, p)) -
-                      (db₃dx₃(t, q) * dv₁dq₂(t, q, p) + db₃dx₂(t, q) * dv₁dq₃(t, q, p))
-
-
-d²g₂dq₃dq₁(t, q, p) = (d²b₁dx₁dx₃(t, q) * v₃(t, q, p) + b₁(t, q) * d²v₃dq₁dq₃(t, q, p)) +
-                      (d²b₁dx₁dx₃(t, q) * v₃(t, q, p) + b₁(t, q) * d²v₃dq₁dq₃(t, q, p)) -
-                      (db₃dx₁(t, q) * dv₁dq₃(t, q, p) + db₃dx₃(t, q) * dv₁dq₁(t, q, p)) -
-                      (db₃dx₁(t, q) * dv₁dq₃(t, q, p) + db₃dx₃(t, q) * dv₁dq₁(t, q, p))
-
-d²g₂dq₃dq₂(t, q, p) = (d²b₁dx₂dx₃(t, q) * v₃(t, q, p) + b₁(t, q) * d²v₃dq₂dq₃(t, q, p)) +
-                      (d²b₁dx₂dx₃(t, q) * v₃(t, q, p) + b₁(t, q) * d²v₃dq₂dq₃(t, q, p)) -
-                      (db₃dx₂(t, q) * dv₁dq₃(t, q, p) + db₃dx₃(t, q) * dv₁dq₂(t, q, p)) -
-                      (db₃dx₂(t, q) * dv₁dq₃(t, q, p) + db₃dx₃(t, q) * dv₁dq₂(t, q, p))
-
-d²g₂dq₃dq₃(t, q, p) = (d²b₁dx₃dx₃(t, q) * v₃(t, q, p) + b₁(t, q) * d²v₃dq₃dq₃(t, q, p)) +
-                      (d²b₁dx₃dx₃(t, q) * v₃(t, q, p) + b₁(t, q) * d²v₃dq₃dq₃(t, q, p)) -
-                      (db₃dx₃(t, q) * dv₁dq₃(t, q, p) + db₃dx₃(t, q) * dv₁dq₃(t, q, p)) -
-                      (db₃dx₃(t, q) * dv₁dq₃(t, q, p) + db₃dx₃(t, q) * dv₁dq₃(t, q, p))
-
-
-d²g₁dq₁dp₁(t, q, p) = db₁dx₁(t, q) * dv₂dp₁(t, q, p) - db₂dx₁(t, q) * dv₁dp₁(t, q, p)
-d²g₁dq₁dp₂(t, q, p) = db₁dx₁(t, q) * dv₂dp₂(t, q, p) - db₂dx₁(t, q) * dv₁dp₂(t, q, p)
-d²g₁dq₁dp₃(t, q, p) = db₁dx₁(t, q) * dv₂dp₃(t, q, p) - db₂dx₁(t, q) * dv₁dp₃(t, q, p)
-
-d²g₁dq₂dp₁(t, q, p) = db₁dx₂(t, q) * dv₂dp₁(t, q, p) - db₂dx₂(t, q) * dv₁dp₁(t, q, p)
-d²g₁dq₂dp₂(t, q, p) = db₁dx₂(t, q) * dv₂dp₂(t, q, p) - db₂dx₂(t, q) * dv₁dp₂(t, q, p)
-d²g₁dq₂dp₃(t, q, p) = db₁dx₂(t, q) * dv₂dp₃(t, q, p) - db₂dx₂(t, q) * dv₁dp₃(t, q, p)
-
-d²g₁dq₃dp₁(t, q, p) = db₁dx₃(t, q) * dv₂dp₁(t, q, p) - db₂dx₃(t, q) * dv₁dp₁(t, q, p)
-d²g₁dq₃dp₂(t, q, p) = db₁dx₃(t, q) * dv₂dp₂(t, q, p) - db₂dx₃(t, q) * dv₁dp₂(t, q, p)
-d²g₁dq₃dp₃(t, q, p) = db₁dx₃(t, q) * dv₂dp₃(t, q, p) - db₂dx₃(t, q) * dv₁dp₃(t, q, p)
-
-
-d²g₂dq₁dp₁(t, q, p) = db₁dx₁(t, q) * dv₃dp₁(t, q, p) - db₃dx₁(t, q) * dv₁dp₁(t, q, p)
-d²g₂dq₁dp₂(t, q, p) = db₁dx₁(t, q) * dv₃dp₂(t, q, p) - db₃dx₁(t, q) * dv₁dp₂(t, q, p)
-d²g₂dq₁dp₃(t, q, p) = db₁dx₁(t, q) * dv₃dp₃(t, q, p) - db₃dx₁(t, q) * dv₁dp₃(t, q, p)
-
-d²g₂dq₂dp₁(t, q, p) = db₁dx₂(t, q) * dv₃dp₁(t, q, p) - db₃dx₂(t, q) * dv₁dp₁(t, q, p)
-d²g₂dq₂dp₂(t, q, p) = db₁dx₂(t, q) * dv₃dp₂(t, q, p) - db₃dx₂(t, q) * dv₁dp₂(t, q, p)
-d²g₂dq₂dp₃(t, q, p) = db₁dx₂(t, q) * dv₃dp₃(t, q, p) - db₃dx₂(t, q) * dv₁dp₃(t, q, p)
-
-d²g₂dq₃dp₁(t, q, p) = db₁dx₃(t, q) * dv₃dp₁(t, q, p) - db₃dx₃(t, q) * dv₁dp₁(t, q, p)
-d²g₂dq₃dp₂(t, q, p) = db₁dx₃(t, q) * dv₃dp₂(t, q, p) - db₃dx₃(t, q) * dv₁dp₂(t, q, p)
-d²g₂dq₃dp₃(t, q, p) = db₁dx₃(t, q) * dv₃dp₃(t, q, p) - db₃dx₃(t, q) * dv₁dp₃(t, q, p)
-
-
-d²g₁dp₁dp₁(t, q, p) = zero(eltype(q))
-d²g₁dp₁dp₂(t, q, p) = zero(eltype(q))
-d²g₁dp₁dp₃(t, q, p) = zero(eltype(q))
-
-d²g₁dp₂dp₁(t, q, p) = zero(eltype(q))
-d²g₁dp₂dp₂(t, q, p) = zero(eltype(q))
-d²g₁dp₂dp₃(t, q, p) = zero(eltype(q))
-
-d²g₁dp₃dp₁(t, q, p) = zero(eltype(q))
-d²g₁dp₃dp₂(t, q, p) = zero(eltype(q))
-d²g₁dp₃dp₃(t, q, p) = zero(eltype(q))
-
-
-d²g₂dp₁dp₁(t, q, p) = zero(eltype(q))
-d²g₂dp₁dp₂(t, q, p) = zero(eltype(q))
-d²g₂dp₁dp₃(t, q, p) = zero(eltype(q))
-
-d²g₂dp₂dp₁(t, q, p) = zero(eltype(q))
-d²g₂dp₂dp₂(t, q, p) = zero(eltype(q))
-d²g₂dp₂dp₃(t, q, p) = zero(eltype(q))
-
-d²g₂dp₃dp₁(t, q, p) = zero(eltype(q))
-d²g₂dp₃dp₂(t, q, p) = zero(eltype(q))
-d²g₂dp₃dp₃(t, q, p) = zero(eltype(q))
+# dg₂dp₁(t, q, p) = b₂(t, q) * dv₃dp₁(t, q, p) - b₃(t, q) * dv₂dp₁(t, q, p)
+# dg₂dp₂(t, q, p) = b₂(t, q) * dv₃dp₂(t, q, p) - b₃(t, q) * dv₂dp₂(t, q, p)
+# dg₂dp₃(t, q, p) = b₂(t, q) * dv₃dp₃(t, q, p) - b₃(t, q) * dv₂dp₃(t, q, p)
 
 
 λₒ(t, q, p) = (
@@ -750,275 +234,30 @@ d²g₂dp₃dp₃(t, q, p) = zero(eltype(q))
     dg₁dp₃(t, q, p) * dHdq₃(t, q, p, params)
 ) / λₒ(t, q, p)
 
-dλₒdq₁(t, q, p) = (
-    d²g₁dq₁dq₁(t, q, p) * dg₂dp₁(t, q, p) +
-    d²g₁dq₁dq₂(t, q, p) * dg₂dp₂(t, q, p) +
-    d²g₁dq₁dq₃(t, q, p) * dg₂dp₃(t, q, p) -
-    d²g₁dq₁dp₁(t, q, p) * dg₂dq₁(t, q, p) -
-    d²g₁dq₁dp₂(t, q, p) * dg₂dq₂(t, q, p) -
-    d²g₁dq₁dp₃(t, q, p) * dg₂dq₃(t, q, p) +
-    dg₁dq₁(t, q, p) * d²g₂dq₁dp₁(t, q, p) +
-    dg₁dq₂(t, q, p) * d²g₂dq₁dp₂(t, q, p) +
-    dg₁dq₃(t, q, p) * d²g₂dq₁dp₃(t, q, p) -
-    dg₁dp₁(t, q, p) * d²g₂dq₁dq₁(t, q, p) -
-    dg₁dp₂(t, q, p) * d²g₂dq₁dq₂(t, q, p) -
-    dg₁dp₃(t, q, p) * d²g₂dq₁dq₃(t, q, p)
-)
 
-dλₒdq₂(t, q, p) = (
-    d²g₁dq₁dq₂(t, q, p) * dg₂dp₁(t, q, p) +
-    d²g₁dq₂dq₂(t, q, p) * dg₂dp₂(t, q, p) +
-    d²g₁dq₂dq₃(t, q, p) * dg₂dp₃(t, q, p) -
-    d²g₁dq₂dp₁(t, q, p) * dg₂dq₁(t, q, p) -
-    d²g₁dq₂dp₂(t, q, p) * dg₂dq₂(t, q, p) -
-    d²g₁dq₂dp₃(t, q, p) * dg₂dq₃(t, q, p) +
-    dg₁dq₁(t, q, p) * d²g₂dq₂dp₁(t, q, p) +
-    dg₁dq₂(t, q, p) * d²g₂dq₂dp₂(t, q, p) +
-    dg₁dq₃(t, q, p) * d²g₂dq₂dp₃(t, q, p) -
-    dg₁dp₁(t, q, p) * d²g₂dq₁dq₂(t, q, p) -
-    dg₁dp₂(t, q, p) * d²g₂dq₂dq₂(t, q, p) -
-    dg₁dp₃(t, q, p) * d²g₂dq₂dq₃(t, q, p)
-)
 
-dλₒdq₃(t, q, p) = (
-    d²g₁dq₁dq₃(t, q, p) * dg₂dp₁(t, q, p) +
-    d²g₁dq₂dq₃(t, q, p) * dg₂dp₂(t, q, p) +
-    d²g₁dq₃dq₃(t, q, p) * dg₂dp₃(t, q, p) -
-    d²g₁dq₃dp₁(t, q, p) * dg₂dq₁(t, q, p) -
-    d²g₁dq₃dp₂(t, q, p) * dg₂dq₂(t, q, p) -
-    d²g₁dq₃dp₃(t, q, p) * dg₂dq₃(t, q, p) +
-    dg₁dq₁(t, q, p) * d²g₂dq₃dp₁(t, q, p) +
-    dg₁dq₂(t, q, p) * d²g₂dq₃dp₂(t, q, p) +
-    dg₁dq₃(t, q, p) * d²g₂dq₃dp₃(t, q, p) -
-    dg₁dp₁(t, q, p) * d²g₂dq₁dq₃(t, q, p) -
-    dg₁dp₂(t, q, p) * d²g₂dq₂dq₃(t, q, p) -
-    dg₁dp₃(t, q, p) * d²g₂dq₃dq₃(t, q, p)
-)
+# function guiding_center_3d_ode_v(v, t, x, params)
+#     q = @views x[1:3]
+#     p = @views x[4:6]
 
-dλₒdp₁(t, q, p) = (
-    d²g₁dq₁dp₁(t, q, p) * dg₂dp₁(t, q, p) +
-    d²g₁dq₂dp₁(t, q, p) * dg₂dp₂(t, q, p) +
-    d²g₁dq₃dp₁(t, q, p) * dg₂dp₃(t, q, p) -
-    d²g₁dp₁dp₁(t, q, p) * dg₂dq₁(t, q, p) -
-    d²g₁dp₁dp₂(t, q, p) * dg₂dq₂(t, q, p) -
-    d²g₁dp₁dp₃(t, q, p) * dg₂dq₃(t, q, p) +
-    dg₁dq₁(t, q, p) * d²g₂dp₁dp₁(t, q, p) +
-    dg₁dq₂(t, q, p) * d²g₂dp₁dp₂(t, q, p) +
-    dg₁dq₃(t, q, p) * d²g₂dp₁dp₃(t, q, p) -
-    dg₁dp₁(t, q, p) * d²g₂dq₁dp₁(t, q, p) -
-    dg₁dp₂(t, q, p) * d²g₂dq₂dp₁(t, q, p) -
-    dg₁dp₃(t, q, p) * d²g₂dq₃dp₁(t, q, p)
-)
+#     v[1] = dHdp₁(t, q, p, params) + λ₁(t, q, p, params) * dg₁dp₁(t, q, p) + λ₂(t, q, p, params) * dg₂dp₁(t, q, p)
+#     v[2] = dHdp₂(t, q, p, params) + λ₁(t, q, p, params) * dg₁dp₂(t, q, p) + λ₂(t, q, p, params) * dg₂dp₂(t, q, p)
+#     v[3] = dHdp₃(t, q, p, params) + λ₁(t, q, p, params) * dg₁dp₃(t, q, p) + λ₂(t, q, p, params) * dg₂dp₃(t, q, p)
+#     v[4] = -dHdq₁(t, q, p, params) - λ₁(t, q, p, params) * dg₁dq₁(t, q, p) - λ₂(t, q, p, params) * dg₂dq₁(t, q, p)
+#     v[5] = -dHdq₂(t, q, p, params) - λ₁(t, q, p, params) * dg₁dq₂(t, q, p) - λ₂(t, q, p, params) * dg₂dq₂(t, q, p)
+#     v[6] = -dHdq₃(t, q, p, params) - λ₁(t, q, p, params) * dg₁dq₃(t, q, p) - λ₂(t, q, p, params) * dg₂dq₃(t, q, p)
 
-dλₒdp₂(t, q, p) = (
-    d²g₁dq₁dp₂(t, q, p) * dg₂dp₁(t, q, p) +
-    d²g₁dq₂dp₂(t, q, p) * dg₂dp₂(t, q, p) +
-    d²g₁dq₃dp₂(t, q, p) * dg₂dp₃(t, q, p) -
-    d²g₁dp₁dp₂(t, q, p) * dg₂dq₁(t, q, p) -
-    d²g₁dp₂dp₂(t, q, p) * dg₂dq₂(t, q, p) -
-    d²g₁dp₂dp₃(t, q, p) * dg₂dq₃(t, q, p) +
-    dg₁dq₁(t, q, p) * d²g₂dp₁dp₂(t, q, p) +
-    dg₁dq₂(t, q, p) * d²g₂dp₂dp₂(t, q, p) +
-    dg₁dq₃(t, q, p) * d²g₂dp₂dp₃(t, q, p) -
-    dg₁dp₁(t, q, p) * d²g₂dq₁dp₂(t, q, p) -
-    dg₁dp₂(t, q, p) * d²g₂dq₂dp₂(t, q, p) -
-    dg₁dp₃(t, q, p) * d²g₂dq₃dp₂(t, q, p)
-)
+#     nothing
+# end
 
-dλₒdp₃(t, q, p) = (
-    d²g₁dq₁dp₃(t, q, p) * dg₂dp₁(t, q, p) +
-    d²g₁dq₂dp₃(t, q, p) * dg₂dp₂(t, q, p) +
-    d²g₁dq₃dp₃(t, q, p) * dg₂dp₃(t, q, p) -
-    d²g₁dp₁dp₃(t, q, p) * dg₂dq₁(t, q, p) -
-    d²g₁dp₂dp₃(t, q, p) * dg₂dq₂(t, q, p) -
-    d²g₁dp₃dp₃(t, q, p) * dg₂dq₃(t, q, p) +
-    dg₁dq₁(t, q, p) * d²g₂dp₁dp₃(t, q, p) +
-    dg₁dq₂(t, q, p) * d²g₂dp₂dp₃(t, q, p) +
-    dg₁dq₃(t, q, p) * d²g₂dp₃dp₃(t, q, p) -
-    dg₁dp₁(t, q, p) * d²g₂dq₁dp₃(t, q, p) -
-    dg₁dp₂(t, q, p) * d²g₂dq₂dp₃(t, q, p) -
-    dg₁dp₃(t, q, p) * d²g₂dq₃dp₃(t, q, p)
-)
-
-dλ₁dq₁(t, q, p, params) = -dλₒdq₁(t, q, p) * λ₁(t, q, p, params) / λₒ(t, q, p) + (
-    d²g₂dq₁dq₁(t, q, p) * dHdp₁(t, q, p, params) +
-    d²g₂dq₁dq₂(t, q, p) * dHdp₂(t, q, p, params) +
-    d²g₂dq₁dq₃(t, q, p) * dHdp₃(t, q, p, params) -
-    d²g₂dq₁dp₁(t, q, p) * dHdq₁(t, q, p, params) -
-    d²g₂dq₁dp₂(t, q, p) * dHdq₂(t, q, p, params) -
-    d²g₂dq₁dp₃(t, q, p) * dHdq₃(t, q, p, params) +
-    dg₂dq₁(t, q, p) * d²Hdq₁dp₁(t, q, p, params) +
-    dg₂dq₂(t, q, p) * d²Hdq₁dp₂(t, q, p, params) +
-    dg₂dq₃(t, q, p) * d²Hdq₁dp₃(t, q, p, params) -
-    dg₂dp₁(t, q, p) * d²Hdq₁dq₁(t, q, p, params) -
-    dg₂dp₂(t, q, p) * d²Hdq₁dq₂(t, q, p, params) -
-    dg₂dp₃(t, q, p) * d²Hdq₁dq₃(t, q, p, params)
-) / λₒ(t, q, p)
-
-dλ₁dq₂(t, q, p, params) = -dλₒdq₂(t, q, p) * λ₁(t, q, p, params) / λₒ(t, q, p) + (
-    d²g₂dq₁dq₂(t, q, p) * dHdp₁(t, q, p, params) +
-    d²g₂dq₂dq₂(t, q, p) * dHdp₂(t, q, p, params) +
-    d²g₂dq₂dq₃(t, q, p) * dHdp₃(t, q, p, params) -
-    d²g₂dq₂dp₁(t, q, p) * dHdq₁(t, q, p, params) -
-    d²g₂dq₂dp₂(t, q, p) * dHdq₂(t, q, p, params) -
-    d²g₂dq₂dp₃(t, q, p) * dHdq₃(t, q, p, params) +
-    dg₂dq₁(t, q, p) * d²Hdq₂dp₁(t, q, p, params) +
-    dg₂dq₂(t, q, p) * d²Hdq₂dp₂(t, q, p, params) +
-    dg₂dq₃(t, q, p) * d²Hdq₂dp₃(t, q, p, params) -
-    dg₂dp₁(t, q, p) * d²Hdq₁dq₂(t, q, p, params) -
-    dg₂dp₂(t, q, p) * d²Hdq₂dq₂(t, q, p, params) -
-    dg₂dp₃(t, q, p) * d²Hdq₂dq₃(t, q, p, params)
-) / λₒ(t, q, p)
-
-dλ₁dq₃(t, q, p, params) = -dλₒdq₃(t, q, p) * λ₁(t, q, p, params) / λₒ(t, q, p) + (
-    d²g₂dq₁dq₃(t, q, p) * dHdp₁(t, q, p, params) +
-    d²g₂dq₂dq₃(t, q, p) * dHdp₂(t, q, p, params) +
-    d²g₂dq₃dq₃(t, q, p) * dHdp₃(t, q, p, params) -
-    d²g₂dq₃dp₁(t, q, p) * dHdq₁(t, q, p, params) -
-    d²g₂dq₃dp₂(t, q, p) * dHdq₂(t, q, p, params) -
-    d²g₂dq₃dp₃(t, q, p) * dHdq₃(t, q, p, params) +
-    dg₂dq₁(t, q, p) * d²Hdq₃dp₁(t, q, p, params) +
-    dg₂dq₂(t, q, p) * d²Hdq₃dp₂(t, q, p, params) +
-    dg₂dq₃(t, q, p) * d²Hdq₃dp₃(t, q, p, params) -
-    dg₂dp₁(t, q, p) * d²Hdq₁dq₃(t, q, p, params) -
-    dg₂dp₂(t, q, p) * d²Hdq₂dq₃(t, q, p, params) -
-    dg₂dp₃(t, q, p) * d²Hdq₃dq₃(t, q, p, params)
-) / λₒ(t, q, p)
-
-dλ₁dp₁(t, q, p, params) = -dλₒdp₁(t, q, p) * λ₁(t, q, p, params) / λₒ(t, q, p) + (
-    d²g₂dq₁dp₁(t, q, p) * dHdp₁(t, q, p, params) +
-    d²g₂dq₂dp₁(t, q, p) * dHdp₂(t, q, p, params) +
-    d²g₂dq₃dp₁(t, q, p) * dHdp₃(t, q, p, params) -
-    d²g₂dp₁dp₁(t, q, p) * dHdq₁(t, q, p, params) -
-    d²g₂dp₁dp₂(t, q, p) * dHdq₂(t, q, p, params) -
-    d²g₂dp₁dp₃(t, q, p) * dHdq₃(t, q, p, params) +
-    dg₂dq₁(t, q, p) * d²Hdp₁dp₁(t, q, p, params) +
-    dg₂dq₂(t, q, p) * d²Hdp₁dp₂(t, q, p, params) +
-    dg₂dq₃(t, q, p) * d²Hdp₁dp₃(t, q, p, params) -
-    dg₂dp₁(t, q, p) * d²Hdq₁dp₁(t, q, p, params) -
-    dg₂dp₂(t, q, p) * d²Hdq₂dp₁(t, q, p, params) -
-    dg₂dp₃(t, q, p) * d²Hdq₃dp₁(t, q, p, params)
-) / λₒ(t, q, p)
-
-dλ₁dp₂(t, q, p, params) = -dλₒdp₂(t, q, p) * λ₁(t, q, p, params) / λₒ(t, q, p) + (
-    d²g₂dq₁dp₂(t, q, p) * dHdp₁(t, q, p, params) +
-    d²g₂dq₂dp₂(t, q, p) * dHdp₂(t, q, p, params) +
-    d²g₂dq₃dp₂(t, q, p) * dHdp₃(t, q, p, params) -
-    d²g₂dp₁dp₂(t, q, p) * dHdq₁(t, q, p, params) -
-    d²g₂dp₂dp₂(t, q, p) * dHdq₂(t, q, p, params) -
-    d²g₂dp₂dp₃(t, q, p) * dHdq₃(t, q, p, params) +
-    dg₂dq₁(t, q, p) * d²Hdp₁dp₂(t, q, p, params) +
-    dg₂dq₂(t, q, p) * d²Hdp₂dp₂(t, q, p, params) +
-    dg₂dq₃(t, q, p) * d²Hdp₂dp₃(t, q, p, params) -
-    dg₂dp₁(t, q, p) * d²Hdq₁dp₂(t, q, p, params) -
-    dg₂dp₂(t, q, p) * d²Hdq₂dp₂(t, q, p, params) -
-    dg₂dp₃(t, q, p) * d²Hdq₃dp₂(t, q, p, params)
-) / λₒ(t, q, p)
-
-dλ₁dp₃(t, q, p, params) = -dλₒdp₃(t, q, p) * λ₁(t, q, p, params) / λₒ(t, q, p) + (
-    d²g₂dq₁dp₃(t, q, p) * dHdp₁(t, q, p, params) +
-    d²g₂dq₂dp₃(t, q, p) * dHdp₂(t, q, p, params) +
-    d²g₂dq₃dp₃(t, q, p) * dHdp₃(t, q, p, params) -
-    d²g₂dp₁dp₃(t, q, p) * dHdq₁(t, q, p, params) -
-    d²g₂dp₂dp₃(t, q, p) * dHdq₂(t, q, p, params) -
-    d²g₂dp₃dp₃(t, q, p) * dHdq₃(t, q, p, params) +
-    dg₂dq₁(t, q, p) * d²Hdp₁dp₃(t, q, p, params) +
-    dg₂dq₂(t, q, p) * d²Hdp₂dp₃(t, q, p, params) +
-    dg₂dq₃(t, q, p) * d²Hdp₃dp₃(t, q, p, params) -
-    dg₂dp₁(t, q, p) * d²Hdq₁dp₃(t, q, p, params) -
-    dg₂dp₂(t, q, p) * d²Hdq₂dp₃(t, q, p, params) -
-    dg₂dp₃(t, q, p) * d²Hdq₃dp₃(t, q, p, params)
-) / λₒ(t, q, p)
-
-dλ₂dq₁(t, q, p, params) = +dλₒdq₁(t, q, p) * λ₂(t, q, p, params) / λₒ(t, q, p) - (
-    d²g₁dq₁dq₁(t, q, p) * dHdp₁(t, q, p, params) +
-    d²g₁dq₁dq₂(t, q, p) * dHdp₂(t, q, p, params) +
-    d²g₁dq₁dq₃(t, q, p) * dHdp₃(t, q, p, params) -
-    d²g₁dq₁dp₁(t, q, p) * dHdq₁(t, q, p, params) -
-    d²g₁dq₁dp₂(t, q, p) * dHdq₂(t, q, p, params) -
-    d²g₁dq₁dp₃(t, q, p) * dHdq₃(t, q, p, params) +
-    dg₁dq₁(t, q, p) * d²Hdq₁dp₁(t, q, p, params) +
-    dg₁dq₂(t, q, p) * d²Hdq₁dp₂(t, q, p, params) +
-    dg₁dq₃(t, q, p) * d²Hdq₁dp₃(t, q, p, params) -
-    dg₁dp₁(t, q, p) * d²Hdq₁dq₁(t, q, p, params) -
-    dg₁dp₂(t, q, p) * d²Hdq₁dq₂(t, q, p, params) -
-    dg₁dp₃(t, q, p) * d²Hdq₁dq₃(t, q, p, params)
-) / λₒ(t, q, p)
-
-dλ₂dq₂(t, q, p, params) = +dλₒdq₂(t, q, p) * λ₂(t, q, p, params) / λₒ(t, q, p) - (
-    d²g₁dq₁dq₂(t, q, p) * dHdp₁(t, q, p, params) +
-    d²g₁dq₂dq₂(t, q, p) * dHdp₂(t, q, p, params) +
-    d²g₁dq₂dq₃(t, q, p) * dHdp₃(t, q, p, params) -
-    d²g₁dq₂dp₁(t, q, p) * dHdq₁(t, q, p, params) -
-    d²g₁dq₂dp₂(t, q, p) * dHdq₂(t, q, p, params) -
-    d²g₁dq₂dp₃(t, q, p) * dHdq₃(t, q, p, params) +
-    dg₁dq₁(t, q, p) * d²Hdq₂dp₁(t, q, p, params) +
-    dg₁dq₂(t, q, p) * d²Hdq₂dp₂(t, q, p, params) +
-    dg₁dq₃(t, q, p) * d²Hdq₂dp₃(t, q, p, params) -
-    dg₁dp₁(t, q, p) * d²Hdq₁dq₂(t, q, p, params) -
-    dg₁dp₂(t, q, p) * d²Hdq₂dq₂(t, q, p, params) -
-    dg₁dp₃(t, q, p) * d²Hdq₂dq₃(t, q, p, params)
-) / λₒ(t, q, p)
-
-dλ₂dq₃(t, q, p, params) = +dλₒdq₃(t, q, p) * λ₂(t, q, p, params) / λₒ(t, q, p) - (
-    d²g₁dq₁dq₃(t, q, p) * dHdp₁(t, q, p, params) +
-    d²g₁dq₂dq₃(t, q, p) * dHdp₂(t, q, p, params) +
-    d²g₁dq₃dq₃(t, q, p) * dHdp₃(t, q, p, params) -
-    d²g₁dq₃dp₁(t, q, p) * dHdq₁(t, q, p, params) -
-    d²g₁dq₃dp₂(t, q, p) * dHdq₂(t, q, p, params) -
-    d²g₁dq₃dp₃(t, q, p) * dHdq₃(t, q, p, params) +
-    dg₁dq₁(t, q, p) * d²Hdq₃dp₁(t, q, p, params) +
-    dg₁dq₂(t, q, p) * d²Hdq₃dp₂(t, q, p, params) +
-    dg₁dq₃(t, q, p) * d²Hdq₃dp₃(t, q, p, params) -
-    dg₁dp₁(t, q, p) * d²Hdq₁dq₃(t, q, p, params) -
-    dg₁dp₂(t, q, p) * d²Hdq₂dq₃(t, q, p, params) -
-    dg₁dp₃(t, q, p) * d²Hdq₃dq₃(t, q, p, params)
-) / λₒ(t, q, p)
-
-dλ₂dp₁(t, q, p, params) = +dλₒdp₁(t, q, p) * λ₂(t, q, p, params) / λₒ(t, q, p) - (
-    d²g₁dq₁dp₁(t, q, p) * dHdp₁(t, q, p, params) +
-    d²g₁dq₂dp₁(t, q, p) * dHdp₂(t, q, p, params) +
-    d²g₁dq₃dp₁(t, q, p) * dHdp₃(t, q, p, params) -
-    d²g₁dp₁dp₁(t, q, p) * dHdq₁(t, q, p, params) -
-    d²g₁dp₁dp₂(t, q, p) * dHdq₂(t, q, p, params) -
-    d²g₁dp₁dp₃(t, q, p) * dHdq₃(t, q, p, params) +
-    dg₁dq₁(t, q, p) * d²Hdp₁dp₁(t, q, p, params) +
-    dg₁dq₂(t, q, p) * d²Hdp₁dp₂(t, q, p, params) +
-    dg₁dq₃(t, q, p) * d²Hdp₁dp₃(t, q, p, params) -
-    dg₁dp₁(t, q, p) * d²Hdq₁dp₁(t, q, p, params) -
-    dg₁dp₂(t, q, p) * d²Hdq₂dp₁(t, q, p, params) -
-    dg₁dp₃(t, q, p) * d²Hdq₃dp₁(t, q, p, params)
-) / λₒ(t, q, p)
-
-dλ₂dp₂(t, q, p, params) = +dλₒdp₂(t, q, p) * λ₂(t, q, p, params) / λₒ(t, q, p) - (
-    d²g₁dq₁dp₂(t, q, p) * dHdp₁(t, q, p, params) +
-    d²g₁dq₂dp₂(t, q, p) * dHdp₂(t, q, p, params) +
-    d²g₁dq₃dp₂(t, q, p) * dHdp₃(t, q, p, params) -
-    d²g₁dp₁dp₂(t, q, p) * dHdq₁(t, q, p, params) -
-    d²g₁dp₂dp₂(t, q, p) * dHdq₂(t, q, p, params) -
-    d²g₁dp₂dp₃(t, q, p) * dHdq₃(t, q, p, params) +
-    dg₁dq₁(t, q, p) * d²Hdp₁dp₂(t, q, p, params) +
-    dg₁dq₂(t, q, p) * d²Hdp₂dp₂(t, q, p, params) +
-    dg₁dq₃(t, q, p) * d²Hdp₂dp₃(t, q, p, params) -
-    dg₁dp₁(t, q, p) * d²Hdq₁dp₂(t, q, p, params) -
-    dg₁dp₂(t, q, p) * d²Hdq₂dp₂(t, q, p, params) -
-    dg₁dp₃(t, q, p) * d²Hdq₃dp₂(t, q, p, params)
-) / λₒ(t, q, p)
-
-dλ₂dp₃(t, q, p, params) = +dλₒdp₃(t, q, p) * λ₂(t, q, p, params) / λₒ(t, q, p) - (
-    d²g₁dq₁dp₃(t, q, p) * dHdp₁(t, q, p, params) +
-    d²g₁dq₂dp₃(t, q, p) * dHdp₂(t, q, p, params) +
-    d²g₁dq₃dp₃(t, q, p) * dHdp₃(t, q, p, params) -
-    d²g₁dp₁dp₃(t, q, p) * dHdq₁(t, q, p, params) -
-    d²g₁dp₂dp₃(t, q, p) * dHdq₂(t, q, p, params) -
-    d²g₁dp₃dp₃(t, q, p) * dHdq₃(t, q, p, params) +
-    dg₁dq₁(t, q, p) * d²Hdp₁dp₃(t, q, p, params) +
-    dg₁dq₂(t, q, p) * d²Hdp₂dp₃(t, q, p, params) +
-    dg₁dq₃(t, q, p) * d²Hdp₃dp₃(t, q, p, params) -
-    dg₁dp₁(t, q, p) * d²Hdq₁dp₃(t, q, p, params) -
-    dg₁dp₂(t, q, p) * d²Hdq₂dp₃(t, q, p, params) -
-    dg₁dp₃(t, q, p) * d²Hdq₃dp₃(t, q, p, params)
-) / λₒ(t, q, p)
+# function ode(x₀, parameters; tspan=tspan, tstep=Δt, periodic=false)
+#     # println("3D Guiding Center model initial constraints g₁ = $(g₁(t₀, q₀, p₀)) and g₂ = $(g₂(t₀, q₀, p₀))")
+#     ODEProblem(
+#         guiding_center_3d_ode_v,
+#         tspan, tstep, x₀;
+#         parameters=parameters,
+#         periodicity=guiding_center_3d_periodicity(q₀, periodic))
+# end
 
 
 function guiding_center_3d_v(v, t, q, p, params)
@@ -1035,45 +274,6 @@ function guiding_center_3d_f(f, t, q, p, params)
     nothing
 end
 
-function guiding_center_3d_canonical_v(v, t, q, p, params)
-    v[1] = dHdp₁(t, q, p, params) +
-           λ₁(t, q, p, params) * dg₁dp₁(t, q, p) +
-           λ₂(t, q, p, params) * dg₂dp₁(t, q, p) +
-           dλ₁dp₁(t, q, p, params) * g₁(t, q, p) +
-           dλ₂dp₁(t, q, p, params) * g₂(t, q, p)
-    v[2] = dHdp₂(t, q, p, params) +
-           λ₁(t, q, p, params) * dg₁dp₂(t, q, p) +
-           λ₂(t, q, p, params) * dg₂dp₂(t, q, p) +
-           dλ₁dp₂(t, q, p, params) * g₁(t, q, p) +
-           dλ₂dp₂(t, q, p, params) * g₂(t, q, p)
-    v[3] = dHdp₃(t, q, p, params) +
-           λ₁(t, q, p, params) * dg₁dp₃(t, q, p) +
-           λ₂(t, q, p, params) * dg₂dp₃(t, q, p) +
-           dλ₁dp₃(t, q, p, params) * g₁(t, q, p) +
-           dλ₂dp₃(t, q, p, params) * g₂(t, q, p)
-    nothing
-end
-
-function guiding_center_3d_canonical_f(f, t, q, p, params)
-    f[1] = -dHdq₁(t, q, p, params) -
-           λ₁(t, q, p, params) * dg₁dq₁(t, q, p) -
-           λ₂(t, q, p, params) * dg₂dq₁(t, q, p) -
-           dλ₁dq₁(t, q, p, params) * g₁(t, q, p) -
-           dλ₂dq₁(t, q, p, params) * g₂(t, q, p)
-    f[2] = -dHdq₂(t, q, p, params) -
-           λ₁(t, q, p, params) * dg₁dq₂(t, q, p) -
-           λ₂(t, q, p, params) * dg₂dq₂(t, q, p) -
-           dλ₁dq₂(t, q, p, params) * g₁(t, q, p) -
-           dλ₂dq₂(t, q, p, params) * g₂(t, q, p)
-    f[3] = -dHdq₃(t, q, p, params) -
-           λ₁(t, q, p, params) * dg₁dq₃(t, q, p) -
-           λ₂(t, q, p, params) * dg₂dq₃(t, q, p) -
-           dλ₁dq₃(t, q, p, params) * g₁(t, q, p) -
-           dλ₂dq₃(t, q, p, params) * g₂(t, q, p)
-    nothing
-end
-
-
 function hode(q₀, p₀, parameters; tspan=tspan, tstep=Δt, periodic=true)
     # println("3D Guiding Center model initial constraints g₁ = $(g₁(t₀, q₀, p₀)) and g₂ = $(g₂(t₀, q₀, p₀))")
     HODEProblem(
@@ -1081,26 +281,10 @@ function hode(q₀, p₀, parameters; tspan=tspan, tstep=Δt, periodic=true)
         guiding_center_3d_f,
         hamiltonian,
         tspan, tstep, q₀, p₀;
-        # tspan, tstep, q₀, fix_initial_momentum(tspan[begin], q₀, p₀);
         parameters=parameters,
         periodicity=guiding_center_3d_periodicity(q₀, periodic))
 end
 
-function hode(x₀, parameters; kwargs...)
-    hode(initial_conditions(tspan[begin], x₀)..., parameters; kwargs...)
-end
-
-function hode_canonical(q₀, p₀, parameters; tspan=tspan, tstep=Δt, periodic=true)
-    # println("3D Guiding Center model initial constraints g₁ = $(g₁(t₀, q₀, p₀)) and g₂ = $(g₂(t₀, q₀, p₀))")
-    HODEProblem(
-        guiding_center_3d_canonical_v,
-        guiding_center_3d_canonical_f,
-        hamiltonian,
-        tspan, tstep, q₀, p₀;
-        parameters=parameters,
-        periodicity=guiding_center_3d_periodicity(q₀, periodic))
-end
-
-function hode_canonical(x₀, parameters; kwargs...)
-    hode_canonical(initial_conditions(tspan[begin], x₀)..., parameters; kwargs...)
+function hode(x₀, parameters; tspan=tspan, kwargs...)
+    hode(initial_conditions(tspan[begin], x₀)..., parameters; tspan=tspan, kwargs...)
 end
