+++ src/guiding_center_3d/quadratic_potentials.jl
@@ -20,6 +20,7 @@ const tspan = (0.0, 2.5E4)
 initial_conditions_quadratic() = merge(initial_conditions(0.0, [0.3, 0.2, -1.4, 0.3]), (params=(μ=2.5E-3,),))
 
 include("guiding_center_3d_equations.jl")
+include("guiding_center_3d_canonical.jl")
 include("guiding_center_3d_diagnostics.jl")
 
 end
