+++ src/guiding_center_3d/solovev_iter.jl
@@ -21,6 +21,7 @@ const qᵢ = [from_cartesian(0, xᵢ)..., 2.8166280889939737]
 const parameters = (μ=4.607782183567846,)
 
 include("guiding_center_3d_equations.jl")
+include("guiding_center_3d_canonical.jl")
 
 function toroidal_momentum(t, q)
     R(t, q) * ϑ₃(t, q)
