+++ src/guiding_center_3d/tokamak_medium_cylindrical.jl
@@ -55,6 +55,7 @@ function f_surface(s, t)
 end
 
 include("guiding_center_3d_equations.jl")
+include("guiding_center_3d_canonical.jl")
 
 function toroidal_momentum(t, q)
     ϑ₃(t, q)
